requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-docx==1.1.0
PyMuPDF==1.23.8
requests==2.31.0
gunicorn==21.2.0